Add the Secure option
In README.md change the "Select" option to the "Menu" and use -S and -s for the new, "Secure" option.  Change the documentation to explain this. Still have to spell out the configuration names and give examples of their usage. May have to add a section describing the packaged cfg file.